+++ requests_cache/backends/__init__.py
@@ -4,6 +4,7 @@
 from logging import getLogger
 from pathlib import Path
 from typing import Callable, Dict, Iterable, Optional, Type, Union
+from warnings import warn
 
 from .._utils import get_placeholder_class, get_valid_kwargs
 from .base import BaseCache, BaseStorage, DictStorage
@@ -76,6 +77,14 @@ def init_backend(
     if isinstance(backend, BaseCache):
         if cache_name:
             backend.cache_name = str(cache_name)
+            # Database names, file paths, etc. cannot be reliably updated after initialization, so
+            # warn if the user passes both `cache_name` and a backend instance
+            warn(
+                '`cache_name` cannot be set after backend initialization; '
+                'please pass it to the backend class instead',
+                DeprecationWarning,
+                stacklevel=3,
+            )
         return backend
     # If no backend is specified, use SQLite as default, unless the environment doesn't support it
     elif not backend: